fix(signal): Decimal-as-string crash + last-known index fallback
SignalRulesOut used Decimal for threshold fields; Pydantic v2 serialises
Decimal as a JSON string ("19.00"), so vixThreshold/breadthThreshold
arrived as strings and .toFixed() threw TypeError crashing the page.

Fix: change SignalRulesOut thresholds to float (Pydantic coerces Decimal
→ float on model_validate, preserving DB round-trip via SignalRulesUpdate).

Also: add 24-h last-known-good fallback cache for /indices so the Signal
page shows stale market data instead of empty cards when Yahoo is down.
Fix index display names to title-case (Nifty 50/Sensex/…) matching the
SignalPage.tsx lookup, so the Nifty card actually resolves.
Fix remaining outer null guards in MarketSignalCard (!=null vs !==undefined).

+++ backend/dhanradar/signal/schemas.py
@@ -11,9 +11,9 @@ from pydantic import BaseModel, field_validator
 
 
 class SignalRulesOut(BaseModel):
-    nifty_threshold: Decimal
-    vix_threshold: Decimal
-    breadth_threshold: Decimal
+    nifty_threshold: float
+    vix_threshold: float
+    breadth_threshold: float
     deploy_ladder: list[int]
     alerts_on: bool
